pyproject.toml
[project]
name = "juan"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "boto3>=1.38.24",
    "ipykernel>=6.29.5",
    "numpy>=2.2.6",
    "pandas>=2.2.3",
    "pyathena>=3.14.1",
    "python-dotenv>=1.1.0",
    "streamlit>=1.45.1",
]
